New-Leather-Goods-Shop/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
datadump==0.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-admin-interface==0.19.0
django-allauth==0.50.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-yasg==1.20.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipykernel==6.13.0
ipython==8.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.3.0
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysqlclient==2.1.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2022.4.8
platformdirs==2.5.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pwhich==0.1.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pythonanywhere==0.10.1
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schema==0.7.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.9
tk==0.1.0
tornado==6.1
traitlets==5.1.1
typer==0.4.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.0.0
python-dotenv
